move `execute_sql_file` to Utility; remove query.rb
